Move jwt directory to root.
