+++ src/js/App/IDPError/IDPError.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Bullseye, Button, EmptyState, EmptyStateBody, EmptyStateIcon, Title } from '@patternfly/react-core';
 import ExclamationCircleIcon from '@patternfly/react-icons/dist/js/icons/exclamation-circle-icon';
-import { logoutAllTabs } from '../../jwt/jwt';
+import { logoutAllTabs } from '../../../jwt/jwt';
 import { useIntl } from 'react-intl';
 import messages from '../../Messages';
