+++ src/js/auth.ts
@@ -1,13 +1,13 @@
-import { getOfflineToken, wipePostbackParamsThatAreNotForUs } from './jwt/insights/offline';
-import * as jwt from './jwt/jwt';
+import { getOfflineToken, wipePostbackParamsThatAreNotForUs } from '../jwt/offline';
+import * as jwt from '../jwt/jwt';
 import cookie from 'js-cookie';
-import { options as defaultOptions } from './jwt/constants';
+import { options as defaultOptions } from '../jwt/constants';
 import { ACCOUNT_REQUEST_TIMEOUT, ACTIVE_REMOTE_REQUEST, CROSS_ACCESS_ACCOUNT_NUMBER, CROSS_ACCESS_ORG_ID } from './consts';
 import { AxiosResponse } from 'axios';
 
 export type LibJWT = {
   getOfflineToken: () => Promise<AxiosResponse<any>>;
-  jwt: typeof import('./jwt/jwt');
+  jwt: typeof jwt;
   initPromise: Promise<void>;
 };
