+++ src/js/chrome/initialize-jwt.ts
@@ -1,6 +1,6 @@
 import { ChromeUser } from '@redhat-cloud-services/types';
 import { LibJWT } from '../auth';
-import { decodeToken } from '../jwt/jwt';
+import { decodeToken } from '../../jwt/jwt';
 import { spinUpStore } from '../redux-config';
 import { CacheAdapter } from '../../utils/cache';
